Improve stats
