+++ Alfheim/Redux/AppState.swift
@@ -17,6 +17,7 @@ struct AppState {
   var settings = Settings()
   var payment = Payment()
   var catemoji = Catemoji()
+  var statistics: Statistics
   // shared global state
 
   var accountDetail: AccountDetail
@@ -25,6 +26,7 @@ struct AppState {
     let account = Alne.Accounts.expenses
     shared = Shared(account: account, allPayments: [], allTransactions: [])
     accountDetail = AccountDetail(account: account)
+    statistics = Statistics(account: account)
   }
 }
 
@@ -104,100 +106,6 @@ extension AppState {
 
     var amountText: String {
       "\(account.currency.symbol)\(String(format: "%.2f", amount))"
-    }
-
-    var categorizedTransactions: [(String, String, Double)] {
-      categorized
-        .map { category, viewModels in
-          (category, viewModels.first!.catemoji.emoji, viewModels.reduce(0.0, { $0 + $1.amount }))
-        }
-    }
-
-    var categorized: [String: [TransactionViewModel]] {
-      displayTransactions
-        .grouped { $0.catemoji.category.name }
-    }
-
-    var categorizedAmount: [String: Double] {
-      displayTransactions
-        .grouped { $0.catemoji.category.name }
-        .mapValues { $0.reduce(0, { $0 + $1.amount }) }
-    }
-
-    func range(with period: Period) -> Range<Date> {
-      let calendar = Calendar.current
-      let today = calendar.startOfDay(for: Date())
-
-      switch period {
-      case .weekly:
-        let from = calendar.date(byAdding: .day, value: -7, to: today)!
-        return from ..< today
-      case .montly:
-        let week = calendar.date(byAdding: .weekOfYear, value: -7, to: today)!
-        return week.start(of: .week) ..< today
-      case .yearly:
-        let month = calendar.date(byAdding: .month, value: -7, to: today)!
-        return month.start(of: .month) ..< today
-      }
-    }
-
-    func labeledAmount(with period: Period) -> [(String, Double)] {
-      let calendar = Calendar.current
-      let today = calendar.startOfDay(for: Date())
-
-      switch period {
-      case .weekly:
-        let intervals: TimeInterval = 24 * 60 * 60
-        let from = calendar.date(byAdding: .day, value: -6, to: today)!
-        let formatter = DateFormatter()
-        formatter.dateFormat = "E"
-        return stride(from: from, to: Date(), by: intervals)
-          .map { date -> (String, Double) in
-            let label = formatter.string(from: date)
-            let next = date.addingTimeInterval(intervals)
-            let amount = allTransactions.filter {
-              $0.date >= date && $0.date < next
-            }
-            .reduce(0, { $0 + $1.amount })
-            return (label, amount)
-          }
-
-      case .montly:
-        let formatter = DateFormatter()
-        formatter.dateFormat = "dd"
-        var result = [(String, Double)]()
-        for idx in (0..<7).reversed() {
-          let week = calendar.date(byAdding: .weekOfYear, value: -idx, to: Date())!
-          let start = week.start(of: .week)
-          let end = week.end(of: .week)
-
-          let label = formatter.string(from: week)
-          let amount = allTransactions.filter {
-            $0.date >= start && $0.date < end
-          }
-          .reduce(0, { $0 + $1.amount })
-          result.append((label, amount))
-        }
-
-        return result
-      case .yearly:
-        let formatter = DateFormatter()
-        formatter.dateFormat = "MMM"
-        var result = [(String, Double)]()
-        for idx in (0..<7).reversed() {
-          let month = calendar.date(byAdding: .month, value: -idx, to: Date())!
-          let start = month.start(of: .month)
-          let end = month.end(of: .month)
-
-          let label = formatter.string(from: month)
-          let amount = allTransactions.filter {
-            $0.date >= start && $0.date < end
-          }
-          .reduce(0, { $0 + $1.amount })
-          result.append((label, amount))
-        }
-        return result
-      }
     }
   }
 }
